+++ coroutines/src/commonMain/kotlin/com/github/michaelbull/result/coroutines/binding/SuspendableBinding.kt
@@ -1,4 +1,4 @@
-package com.github.michaelbull.result.coroutines
+package com.github.michaelbull.result.coroutines.binding
 
 import com.github.michaelbull.result.Err
 import com.github.michaelbull.result.Ok
